lj_record.c: Trace IFUNCF the same way as FUNCF
When a trace reaches an interpreted (blacklisted) function, continue
tracing into the function, as would happen if the function were not
blacklisted (nor JITed.)

This means that when a function is blacklisted it can never serve as
the start of a trace, but it can still be included in larger traces
that start somewhere else. This can be beneficial because some
functions can be traced perfectly well _except_ in isolation, for
example small utilities that end with a tail-call to a library
function.